Remove redundant evenLevel from Node field
Since this is essentially always known by other means, it's
redundant to also store it within the Node.

For insert() and contains(), we start at root, and we know we're at an
even level.  Whenever we descend, we know we flip the orientation bit.
Accordingly, we always know within insert() and contains() whether we're
on a horizontal or a vertical partition, independent of any field
within Node which may also contain that information.

Accordingly, don't also store orientation information within Node.