Revise KdTree.contains() to break ties consistently
Previously, the assumption was that, if the search point is on some
Node's partition line, then the search point _must_ be equal to the
Node's point, or the BST does not contain the search point.

Unfortunately, this was a mistaken assumption.  Many points can lie
along the same partition line (consider the vertical and horizontal
test files in kdtree-testing).  Even if the current Node's point
does not equal the search point (when that search point is on the
partition line), it's still possible that some other, later-inserted
point will match the search point.

The insert() method breaks ties such as these by placing the point
to-be-added on the right of the current Node.  Accordingly, the
contains() method must also break ties by traversing the right branch
of the current Node.